+++ popkorn/src/main/kotlin/cc/popkorn/annotations/Injectable.kt
@@ -1,10 +1,12 @@
 package cc.popkorn.annotations
 
+import cc.popkorn.Propagation
 import cc.popkorn.Scope
 import kotlin.reflect.KClass
 
 /**
  * Annotation to define that a class can be injected.
+ * By default this annotation adds also as an injectable object all its interfaces (Propagation.ALL) unless otherwise specified
  * Also, if not defined, takes the scope BY_APP by default
  * If you want to exclude some of them, do it through parameter 'exclude'
  *
@@ -16,6 +18,7 @@ import kotlin.reflect.KClass
 annotation class Injectable(
     val scope: Scope = Scope.BY_APP,
     val alias:String = "",
+    val propagation : Propagation = Propagation.ALL,
     vararg val exclude: KClass<*> = []
 )