+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 1.2.0 *(2019-12-XX)*
 -----------------------------
 * New: Free `Injector` from it's internal modifier. Now can be used.
 * New: Now can use parametrized methods when injecting through `InjectableProvider`  
+* New: Add propagation strategy at `Injectable` and `InjectableProvider`
 
 Version 1.1.0 *(2019-11-29)*
 -----------------------------